requirements.txt
# Core
openai==1.12.0
langchain==0.1.9
langchain-openai==0.0.5

# Vector Store
chromadb==0.4.22

# Frontend
streamlit==1.31.0

# Utils
python-dotenv==1.0.0
tiktoken==0.5.2

# Development
pytest==8.0.0
